requirements.txt
xgboost
scikit-learn
pandas
numpy
colorama
tldextract
urlextract
pyspellchecker
oletools
python-magic
pdfid
pefile
email